Ignores minified code
